Move EmptyList into components folder
